third_party/blink/web_tests/external/wpt/app-history/app-history-entry/entries-in-new-srcdoc-iframe.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/is_uuid.js"></script>
<iframe id="i" srcdoc="new"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  assert_not_equals(i.contentWindow.appHistory.current, null);
}, "AppHistory entries() and current should be set in a new srcdoc iframe");
</script>
